pyproject.toml
[tool.poetry]
name = "coronatracker"
version = "0.1.0"
description = ""
authors = ["hoppingturtles <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.6"
Flask = "^2.0.1"
matplotlib = "^3.4.2"
pandas = "^1.2.4"
tqdm = "^4.61.1"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
